CONN-1279a: Updated review comments, deployed WAR and able to access services. Good to merge.
