+++ Product/Production/Gateway/CORE_X12DocumentSubmission_10/src/main/java/gov/hhs/fha/nhinc/corex12/docsubmission/realtime/_10/entity/CORE_EntityX12DSRealTimeUnsecured.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2012, United States Government, as represented by the Secretary of Health and Human Services.
+ * Copyright (c) 2014, United States Government, as represented by the Secretary of Health and Human Services.
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -27,10 +27,10 @@
 package gov.hhs.fha.nhinc.corex12.docsubmission.realtime._10.entity;
 
 import gov.hhs.fha.nhinc.common.nhinccommonentity.RespondingGatewayCrossGatewayBatchSubmissionAckResponseType;
-import gov.hhs.fha.nhinc.common.nhinccommonentity.RespondingGatewayCrossGatewayBatchSubmissionResponseMessageSecuredRequestType;
-import gov.hhs.fha.nhinc.common.nhinccommonentity.RespondingGatewayCrossGatewayBatchSubmissionSecuredRequestType;
+import gov.hhs.fha.nhinc.common.nhinccommonentity.RespondingGatewayCrossGatewayBatchSubmissionRequestType;
+import gov.hhs.fha.nhinc.common.nhinccommonentity.RespondingGatewayCrossGatewayBatchSubmissionResponseMessageRequestType;
+import gov.hhs.fha.nhinc.common.nhinccommonentity.RespondingGatewayCrossGatewayRealTimeRequestType;
 import gov.hhs.fha.nhinc.common.nhinccommonentity.RespondingGatewayCrossGatewayRealTimeResponseType;
-import gov.hhs.fha.nhinc.common.nhinccommonentity.RespondingGatewayCrossGatewayRealTimeSecuredRequestType;
 import gov.hhs.fha.nhinc.corex12.docsubmission.realtime.outbound.OutboundCORE_X12DSRealTime;
 import javax.annotation.Resource;
 import javax.xml.ws.BindingType;
@@ -43,33 +43,56 @@ import javax.xml.ws.soap.Addressing;
  */
 @BindingType(value = javax.xml.ws.soap.SOAPBinding.SOAP12HTTP_BINDING)
 @Addressing(enabled = true)
-public class EntityX12DSRealTimeSecured implements gov.hhs.fha.nhinc.nhincentitycoresecured.EntityCORETransactionSecuredPortType {
+public class CORE_EntityX12DSRealTimeUnsecured implements gov.hhs.fha.nhinc.nhincentitycore.EntityCORETransactionPortType {
 
     private WebServiceContext context;
-    private OutboundCORE_X12DSRealTime outboundCOREX12DSRealTime;
+    private OutboundCORE_X12DSRealTime outboundCORE_X12DSRealTime;
     
+    /**
+     * 
+     * @param body
+     * @return RespondingGatewayCrossGatewayBatchSubmissionAckResponseType
+     */
     @Override
-    public RespondingGatewayCrossGatewayBatchSubmissionAckResponseType batchSubmitTransactionDeferredRequestSecured(RespondingGatewayCrossGatewayBatchSubmissionSecuredRequestType body) {
+    public RespondingGatewayCrossGatewayBatchSubmissionAckResponseType batchSubmitTransactionDeferredResponse(RespondingGatewayCrossGatewayBatchSubmissionResponseMessageRequestType body) {
         throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
     }
 
+    /**
+     * 
+     * @param body
+     * @return RespondingGatewayCrossGatewayBatchSubmissionAckResponseType
+     */
     @Override
-    public RespondingGatewayCrossGatewayRealTimeResponseType realTimeTransactionSecured(RespondingGatewayCrossGatewayRealTimeSecuredRequestType body) {
-        return new EntityX12DSRealTimeImpl(outboundCOREX12DSRealTime).realTimeTransactionSecured(body, context);
-    }
-
-    @Override
-    public RespondingGatewayCrossGatewayBatchSubmissionAckResponseType batchSubmitTransactionDeferredResponseSecured(RespondingGatewayCrossGatewayBatchSubmissionResponseMessageSecuredRequestType body) {
+    public RespondingGatewayCrossGatewayBatchSubmissionAckResponseType batchSubmitTransactionDeferredRequest(RespondingGatewayCrossGatewayBatchSubmissionRequestType body) {
         throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
     }
+
+    /**
+     * 
+     * @param body
+     * @return RespondingGatewayCrossGatewayRealTimeResponseType
+     */
+    @Override
+    public RespondingGatewayCrossGatewayRealTimeResponseType realTimeTransaction(RespondingGatewayCrossGatewayRealTimeRequestType body) {
+        return new CORE_EntityX12DSRealTimeImpl(outboundCORE_X12DSRealTime).realTimeTransaction(body, context);
+    }
     
+    /**
+     * 
+     * @param context 
+     */
     @Resource
     public void setContext(WebServiceContext context) {
         this.context = context;
     }
     
-    public void setOutboundCORE_X12DSRealTime(OutboundCORE_X12DSRealTime outboundCOREX12DSRealTime)
+    /**
+     * 
+     * @param outboundCORE_X12DSRealTime 
+     */
+    public void setOutboundCORE_X12DSRealTime(OutboundCORE_X12DSRealTime outboundCORE_X12DSRealTime)
     {
-        this.outboundCOREX12DSRealTime = outboundCOREX12DSRealTime;
+        this.outboundCORE_X12DSRealTime = outboundCORE_X12DSRealTime;
     }
 }
