+++ Product/Production/Gateway/CORE_X12DocumentSubmission_10/src/main/java/gov/hhs/fha/nhinc/corex12/docsubmission/realtime/_10/nhin/CORE_NhinX12DSRealTime.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2012, United States Government, as represented by the Secretary of Health and Human Services.
+ * Copyright (c) 2014, United States Government, as represented by the Secretary of Health and Human Services.
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -29,7 +29,9 @@ package gov.hhs.fha.nhinc.corex12.docsubmission.realtime._10.nhin;
 
 import gov.hhs.fha.nhinc.corex12.docsubmission.realtime.inbound.InboundCORE_X12DSRealTime;
 import javax.annotation.Resource;
+import javax.xml.ws.BindingType;
 import javax.xml.ws.WebServiceContext;
+import javax.xml.ws.soap.Addressing;
 import org.caqh.soap.wsdl.CORETransactions;
 import org.caqh.soap.wsdl.corerule2_2_0.COREEnvelopeBatchSubmission;
 import org.caqh.soap.wsdl.corerule2_2_0.COREEnvelopeBatchSubmissionAckRetrievalResponse;
@@ -41,37 +43,65 @@ import org.caqh.soap.wsdl.corerule2_2_0.COREEnvelopeRealTimeResponse;
  *
  * @author svalluripalli
  */
-public class NhinX12DSRealTime implements CORETransactions {
+@BindingType(value = javax.xml.ws.soap.SOAPBinding.SOAP12HTTP_BINDING)
+@Addressing(enabled = true)
+public class CORE_NhinX12DSRealTime implements CORETransactions {
 
     private WebServiceContext context;
     private InboundCORE_X12DSRealTime inboundCORE_X12DSRealTime;
     
+    /**
+     * 
+     * @param context 
+     */
     @Resource
     public void setContext(WebServiceContext context) {
         this.context = context;
     }
 
+    /**
+     * 
+     * @return InboundCORE_X12DSRealTime
+     */
     public InboundCORE_X12DSRealTime getInboundCORE_X12DSRealTime() {
         return inboundCORE_X12DSRealTime;
     }
 
+    /**
+     * 
+     * @param inboundCORE_X12DSRealTime 
+     */
     public void setInboundCORE_X12DSRealTime(InboundCORE_X12DSRealTime inboundCORE_X12DSRealTime) {
         this.inboundCORE_X12DSRealTime = inboundCORE_X12DSRealTime;
     }
     
+    /**
+     * 
+     * @param body
+     * @return COREEnvelopeBatchSubmissionAckRetrievalResponse
+     */
     @Override
     public COREEnvelopeBatchSubmissionAckRetrievalResponse batchSubmitTransactionDeferredRequest(COREEnvelopeBatchSubmission body) {
         throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
     }
 
+    /**
+     * 
+     * @param body
+     * @return COREEnvelopeBatchSubmissionAckRetrievalResponse
+     */
     @Override
     public COREEnvelopeBatchSubmissionAckRetrievalResponse batchSubmitTransactionDeferredResponse(COREEnvelopeBatchSubmissionResponse body) {
         throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
     }
 
+    /**
+     * 
+     * @param body
+     * @return COREEnvelopeRealTimeResponse
+     */
     @Override
     public COREEnvelopeRealTimeResponse realTimeTransaction(COREEnvelopeRealTimeRequest body) {
-        return new NhinX12DSRealTimeImpl(inboundCORE_X12DSRealTime).realTimeTransaction(body, context);
-    }
-    
+        return new CORE_NhinX12DSRealTimeImpl(inboundCORE_X12DSRealTime).realTimeTransaction(body, context);
+    }    
 }
