+++ Product/Production/Gateway/CORE_X12DocumentSubmission_10/src/main/webapp/WEB-INF/ibm-web-bnd.xml
@@ -2,6 +2,5 @@
 <web-bnd xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns="http://websphere.ibm.com/xml/ns/javaee"
     xsi:schemaLocation="http://websphere.ibm.com/xml/ns/javaee http://websphere.ibm.com/xml/ns/javaee/ibm-web-bnd_1_0.xsd" version="1.0">
   <resource-ref binding-name="jdbc/transrepo_datasource" name="jdbc/transrepo_datasource"/>
-  <resource-ref binding-name="jdbc/eventdb_datasource" name="jdbc/eventdb_datasource"/>
 </web-bnd>
