+++ Product/Production/Gateway/CORE_X12DocumentSubmission_10/src/main/webapp/WEB-INF/jboss-web.xml
@@ -1,16 +1,8 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<jboss-web>
-    
-       <resource-ref>
+<jboss-web>    
+    <resource-ref>
         <res-ref-name>jdbc/transrepo_datasource</res-ref-name>
         <jndi-name>java:/jdbc/transrepo_datasource</jndi-name>
-        <res-type>javax.sql.DataSource</res-type>
-        <res-auth>Container</res-auth>
-    </resource-ref>
-    
-    <resource-ref>
-        <res-ref-name>jdbc/eventdb_datasource</res-ref-name>
-        <jndi-name>java:/jdbc/eventdb_datasource</jndi-name>
         <res-type>javax.sql.DataSource</res-type>
         <res-auth>Container</res-auth>
     </resource-ref>
