+++ Product/Production/Gateway/CORE_X12DocumentSubmission_10/src/main/webapp/WEB-INF/web.xml
@@ -42,10 +42,4 @@
         <res-auth>Container</res-auth>
         <res-sharing-scope>Shareable</res-sharing-scope>
     </resource-ref>
-    <resource-ref>
-        <res-ref-name>jdbc/eventdb_datasource</res-ref-name>
-        <res-type>javax.sql.DataSource</res-type>
-        <res-auth>Container</res-auth>
-        <res-sharing-scope>Shareable</res-sharing-scope>
-    </resource-ref>
 </web-app>
